+++ packages/apollo-angular/src/index.ts
@@ -12,5 +12,4 @@ export {
   ExtraSubscriptionOptions,
   Flags,
 } from './types';
-export {gql} from '@apollo/client/core';
-export {gql as graphql} from '@apollo/client/core';
+export {gql, graphql} from './gql';